updated script, compressed images

+++ data/scripts/compress_images.sh
@@ -3,15 +3,17 @@ if ! command -v cwebp &> /dev/null; then
     exit 1
 fi
 
-for file in "../../public/images/members"/*.{jpg,png}; do
+for file in "../../public/images/members"/*.{jpg,png,webp}; do
     if [ -f "$file" ]; then
         filename=$(basename -- "$file")
         extension="${filename##*.}"
         filename_no_ext="${filename%.*}"
         out="../../public/images/members/$filename_no_ext.webp"
 
-        cwebp -q 80 "$file" -o "$out"
-        rm "$file"
+        convert "$file" -resize 400x400 "$out"
+        if [ "$extension" != "webp" ]; then
+            rm "$file"
+        fi
         echo "done"
         echo "$filename"
     fi
